initial commit, added first five scripts
More crap on GitHub! Woo! I've added the five scripts I've created so
far, although they aren't all exactly extensive (looking at you,
of-fibonacci.js). The factors.js file is the main one, which is included
by the others. Anything that uses it gets to be part of the repository.

There's a pretty useless (ATM) readme and GPLv3 as per usual.